rename gadget file
