+++ src/lib/provable/gadgets/native-curve.ts
